requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arvix
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai